refactor(blocks): flatten pages/ into blocks/
Move all 10 page block files and the barrel index from src/blocks/pages/
directly into src/blocks/ and delete the pages/ subdirectory. Update the
three importers (Lessons, Pages, content-index-block test) to use @/blocks.

No behavior change — purely structural.

Refs #442
