+++ src/blocks/ContentIndexBlock.ts
